Add unique values support
* 🦄 Add unique values support to Faker

Adds support to Faker for guaranteed unique values.

A new `.unique` property has been added to the Proxy class, which
returns an instance of a `UniqueProxy`. Any attributes looked up on the
unique proxy goes to the common proxy. If the attribute is a function,
then it's wrapped with a wrapper that only allows unique values.

Obviously, it's not always possible to get a unique value, especially
with a small set of values. If the wrapper can't find a unique value
after a certain number of attempts (currently pegged at an arbitrary
1000), it throws a UniquenessSanityException.

Each "remembered" set of values is tied to the name of the function, and
the arguments passed to it, so, for example, different `.random_int()`
calls with different arguments will not share a "bucket".

Since it uses a set internally for remembering previous values, and a
dictionary for determining what bucket to use, it can't be used with any
arguments or return values that are unhashable.

* 🐑 Fix flake8 linting

Trim some unused imports, add some set literals, add some line breaks.
Makes changes as required to pass flake8

* ⚠️ Changed exception name; fallback behaviour

- The exception that the unique wrapper throws after too many iterations
  is now called `UniquenessException`, and also includes an error
  message explaining how many iterations were attempted.

- Accessing a non-callable through the `.unique` attribute now throws a
  TypeError.

* 📦 Fix accidental unpacking of arguments in key

The key for looking up previously generated values of a specific
invocation of a function accidentally unpacked the arguments of `args`,
making key a variable length tuple.

This commit fixes that mistake, so the keys are now consistently
3-tuples.

* 📄 Add documentation for .unique attribute

First pass at documentation for the .unique attribute.

* 📄 Tweak documentation according to feedback

- A version placeholder has been replaced with the version these
  features are expected to go live.

- Some wording on prose has been adjusted.

- Code samples have had some line breaks inserted for readability.

* 🔀 Fix ordering according to isort

* ❓ Change pytest fixture to clear unique values

The fixture clears the generated `.unique` values for the given Faker
instance, to make each test independent of any previous ones.

* 🔀 Fix isort import order

* Update README.rst

* Apply suggestions from code review

* 🚨 Add test for unique fixture clearing

Add a test for checking that using the `faker` fixture across multiple
functions have their record of pre-existing generated unique values
cleared.

* 🐑 Lint and add an extra trailing comma

* ⏭️ Modify plugin test to test successive clears

The test for checking the faker pytest plugin now fully exhausts the
"boolean" bucket, in order to demonstrate that using the unique proxy
does not affect regular execution, and that tests have their "unique"
cache cleared between invocations.

* 🔠 Correct inner test spelling error

"exhaust" is the correct spelling.

Co-authored-by: Flavio Curella <89607+fcurella@users.noreply.github.com>